\name{buildChromMap}
\alias{buildChromMap}
\title{A function to generate an instantiation of a chromLocation class}
\description{
This function will take take the name of a data package and build a chromLocation object representing regions of the genomes.
}
\usage{
buildChromMap(dataPkg, regions)
}

\arguments{
  \item{dataPkg}{The name of the data package to be used (a.k.a generated by AnnBuilder or downloaded from Bioconductor web site}
  \item{regions}{a character vector of genome regions to be generated}
}

\details{
This function is related to the buildChromLocation function found in the 'annotate' library. However, this function can be used to build specialized chromLocation objects based on gene mapping information.  For example, a chromLocation object can be build specifically for human chromosome 1 by supplying chromosomal band information, such as c("1p1", "1p2", "1p3", "1q1", "1q2", "1q3", "1q4").  Genes that map to these regions are isolated and a chromLocation object is returned.
Note that genes are isolated by 'grep'ing genome mapping information.  Therefore the number of genes that are able to placed into a defined genetic region (i.e. 1q4) is dependent on the quality of the mapping information in the annotation data source.

Unfortunately, not too many pre-built annotation packages are available for spotted arrays off the Bioconductor Metadata web set. Use AnnBuilder to make one or get one from your core.  
}


\value{
A 'chromLocation' object representing the specified genomic regions and annotation data source
}


\author{Kyle A. Furge <kyle.furge@vai.org}


\seealso{\code{\link[annotate]{buildChromLocation}}}

\examples{
## 
## NOTE: This requires an annotation package to work.
##       In this example it is hu6800
## 
if (require(hu6800)) {

  library(Biobase)
  library(annotate)

  ## Build a specific chrom arm

  chr1q <- buildChromMap("hu6800",c("1q1","1q2","1q3","1q4"))

  ## Build human data based on chrom arms

  data(Hs.arms) 
  map <- buildChromMap("hu6800",Hs.arms)
}
}
\keyword{manip}
