\name{Hs.arms}
\alias{Hs.arms}
\docType{data}
\title{ Hs.arms }
\description{
  The data set gives the human chromosomal arms.
}
\usage{data(Hs.arms)}
\format{
  The format is:
 chr [1:48] "1p" "1q" "2p" "2q" "3p" "3q" "4p" "4q" "5p" "5q" "6p" "6q" "7p" "7q" "8p" "8q" "9p" ...
}


\source{International System of Human Cytogenetic Nomenclature (ISCN)}



\keyword{datasets}
