\name{mcr.eset}
\alias{mcr.eset}

\docType{data}
\title{ Example exprSet and chromLocation objects}
\description{
  An example exprSet and a chromLocation object generated from an gene
  expression profiling experiment of leukemic and normal blood cells.
  Profiling was done on custom pin-printed cDNA arrays.
}
\usage{data(mcr.eset)}


\source{
  Lindvall C, Furge K, Bjorkholm M, Guo X, Haab B, Blennow E,
  Nordenskjold M, Teh BT. Combined genetic and transcriptional profiling
  of acute myeloid leukemia with normal and complex
  karyotypes. Haematologica. 2004 Sep;89(9):1072-81
  PMID: 15377468
}
\examples{
data(mcr.eset)

str(mcr.eset)

}
\keyword{datasets}
