\name{t.binom.test}
\alias{t.binom.test}
\title{binomial t-test}
\description{
  Binomial t-test
}
\usage{
t.binom.test(x,trim=.1)
}
\arguments{
  \item{x}{numeric argument}
  \item{trim}{trim at?}
}
\value{
  bla bla bla
}
\author{Karl A. Dykema and Kyle A. Furge}
\examples{
  cat("asdf")
}
\keyword{ manip }
