\name{tBinomTest}
\alias{tBinomTest}
\title{binomial t-test}
\description{
  Binomial t-test
}
\usage{
tBinomTest(x,trim=.1)
}
\arguments{
  \item{x}{numeric argument}
  \item{trim}{trim at?}
}
\value{
  bla bla bla
}
\author{Karl A. Dykema and Kyle A. Furge}
\examples{
  cat("this is an example")
}
\keyword{ manip }
